fix x-token typo

+++ api/app/routers/velocity.py
@@ -5,7 +5,7 @@ from app.features.security import verify_token
 router = APIRouter(prefix="/velocity", tags=["velocity"])
 
 class Velocity():
-    def __init__(self, x: int, y: int, z: int):
+    def __init__(self, x: float, y: float, z: float):
         self.x = x
         self.y = y
         self.z = z
